Refactor: Break up fetching logic in more modules - Close #22
